Initial Code commit - Basic resources
'demo' directory covers the deployment of a set of demo components.
It requires an OpenStack environment (as a target for resource build),
but can be modified quickly to use "native"/"baremetal" machines as
targets as well.

demo: ansible to launch a demonstration in an OpenStack or pre-provisioned
compute target.

kolla:  Extensions to the Kolla tools that can make containers
a good choice for OpenStack deployment.